+++ src/algorithm/dp/rainTerraces/dpRainTerraces.ts
@@ -1,21 +1,22 @@
 export function dpRainTerraces(terraces: number[]): number {
   let amount = 0;
+  const dpLeftMax = Array(terraces.length).fill(0);
+  const dpRightMax = Array(terraces.length).fill(0);
 
-  const leftMax = new Array(terraces.length).fill(0);
-  const rightMax = new Array(terraces.length).fill(0);
-
-  let leftTmp = 0;
+  let leftMaxPrevious = 0;
   for (let i = 0; i < terraces.length; i++) {
-    leftTmp = leftMax[i] = Math.max(terraces[i], leftTmp);
+    dpLeftMax[i] = Math.max(terraces[i], leftMaxPrevious);
+    leftMaxPrevious = dpLeftMax[i];
   }
 
-  let rightTmp = 0;
+  let rightMaxPrevious = 0;
   for (let i = terraces.length - 1; i >= 0; i--) {
-    rightTmp = rightMax[i] = Math.max(terraces[i], rightTmp);
+    dpRightMax[i] = Math.max(terraces[i], rightMaxPrevious);
+    rightMaxPrevious = dpRightMax[i];
   }
 
   for (let i = 0; i < terraces.length; i++) {
-    const dx = Math.min(leftMax[i], rightMax[i]) - terraces[i];
+    const dx = Math.min(dpLeftMax[i], dpRightMax[i]) - terraces[i];
     amount += dx > 0 ? dx : 0;
   }