+++ src/sort/countSort/index.ts
@@ -8,19 +8,19 @@ export function countSort(originalArray: number[]): number[] {
   const count = Array(max - min + 1)
     .fill(0);
 
-  originalArray.forEach((item) => {
-    count[item - min]++;
-  });
+  for (let i = 0; i < originalArray.length; i++) {
+    count[originalArray[i] - min]++;
+  }
 
   for (let i = 1; i < count.length; i++) {
     count[i] += count[i - 1];
   }
 
-  const result = Array(originalArray.length);
+  const result = Array(originalArray.length)
+    .fill(null);
   for (let i = originalArray.length - 1; i >= 0; i--) {
     result[count[originalArray[i] - min] - 1] = originalArray[i];
     count[originalArray[i] - min]--;
   }
-
   return result;
 }